mock utils/stores
