+++ frontend/src/util/__mocks__/stores.js
@@ -1,0 +1,15 @@
+module.exports = {
+    useSessionStore: (fn) => {
+        const methods = {
+            setSession: jest.fn(),
+            session: 1
+        } 
+        return fn(methods);
+    },
+    useParticipantStore: () => {
+        return {participant: 1}
+    },
+    useErrorStore: () => {
+        return {setError: jest.fn()}
+    }
+};
